scripts/requirements.txt
# Data Ingestion Scripts Dependencies
# Install with: pip install -r requirements.txt

# Supabase client
supabase>=2.0.0

# HTTP client
httpx>=0.25.0

# Environment variables
python-dotenv>=1.0.0

# Geometry handling
shapely>=2.0.0
geopandas>=0.14.0
pyproj>=3.6.0

# GeoJSON validation
geojson>=3.0.0

# Progress bars
tqdm>=4.66.0

# Data processing
pandas>=2.0.0
numpy>=1.25.0

# Logging
structlog>=23.2.0

# CLI
click>=8.1.0
rich>=13.7.0
